Move locsys test to four_c folder
