+++ tests/testing_meshpy.py
@@ -65,7 +65,6 @@ from meshpy import (
     Beam3rLine2Line2,
     MaterialStVenantKirchhoff,
 )
-from meshpy.four_c.locsys_condition import LocSysCondition
 from meshpy.node import Node, NodeCosserat
 from meshpy.vtk_writer import VTKWriter
 from meshpy.geometry_set import GeometrySet, GeometrySetNodes
@@ -1943,49 +1942,6 @@ class TestMeshpy(unittest.TestCase):
         mpy.import_mesh_full = True
         mesh = self.create_beam_to_solid_conditions_model()
         _plotter = mesh.display_pyvista(is_testing=True, resolution=3)
-
-    def test_meshpy_locsys_condition(self):
-        """
-        Test case for point locsys condition for beams.
-        The testcase is similar to beam3r_herm2line3_static_locsys.dat, but with simpler material.
-        """
-
-        # Create the input file with function and material.
-        input_file = InputFile()
-
-        fun = Function("SYMBOLIC_FUNCTION_OF_SPACE_TIME t")
-        input_file.add(fun)
-
-        mat = MaterialReissner()
-        input_file.add(mat)
-
-        # Create the beam.
-        beam_set = create_beam_mesh_line(
-            input_file, Beam3rHerm2Line3, mat, [2.5, 2.5, 2.5], [4.5, 2.5, 2.5], n_el=1
-        )
-
-        # Add dirichlet boundary conditions.
-        input_file.add(
-            BoundaryCondition(
-                beam_set["start"],
-                "NUMDOF 9 ONOFF 1 1 1 1 1 1 0 0 0 VAL 0.0 0.0 0.0 0.0 0.0 0.0 0.0 0.0 0.0 FUNCT 0 0 0 0 0 0 0 0 0",
-                bc_type=mpy.bc.dirichlet,
-            )
-        )
-
-        input_file.add(
-            BoundaryCondition(
-                beam_set["end"],
-                "NUMDOF 9 ONOFF 1 0 0 0 0 0 0 0 0 VAL 1.0 0.0 0.0 0.0 0.0 0.0 0.0 0.0 0.0 FUNCT 1 0 0 0 0 0 0 0 0",
-                bc_type=mpy.bc.dirichlet,
-            )
-        )
-
-        # Add local sys condition with rotation
-        input_file.add(LocSysCondition(beam_set["end"], Rotation([0, 0, 1], 0.1)))
-
-        # Compare with the reference solution.
-        compare_test_result(self, input_file.get_string(header=False, check_nox=False))
 
 
 if __name__ == "__main__":
